refactor: 헤더 컴포넌트 분리(#87)
