+++ src/features/common/header/Header.tsx
@@ -1,0 +1,13 @@
+import WebHeader from './WebHeader';
+import MobileHeader from './MobileHeader';
+
+function Header() {
+  return (
+    <>
+      <WebHeader />
+      <MobileHeader />
+    </>
+  );
+}
+
+export default Header;
